Add suggestion filtering in JSX and TSX files
Add parsing of JSX and TSX files to only make suggestions in relevant
parts of the document.

Suggestions are only made in JSXText and JSXClosingElement nodes.
Including JSXClosingElement enables suggestions in empty tags (for
example in the middle of <div></div>).
No suggestions are made when parsing fails, avoiding suggestions
resulting in invalid syntax.